Change num1 input to integer type

+++ sample.py
@@ -1,5 +1,5 @@
 # Prompt the user for two numbers
-num1 = float(input("Enter the first number: "))
+num1 = int(input("Enter the first number: "))
 num2 = float(input("Enter the second number: "))
 
 # Perform calculations
